+++ actix-files/CHANGES.md
@@ -1,7 +1,7 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
-- `Files`: `%2F` in request URL path is now decoded to `/` and thus functions as a path separator. [#2398]
+- `Files`: request URL paths with `%2F` are now rejected. [#2398]
 - `Files`: Fixed a regression where `%25` in the URL path is not decoded to `%` in the file path. [#2398]
 - Minimum supported Rust version (MSRV) is now 1.54.